remove unintended implementations
