+++ src/Utils/request/types.ts
@@ -23,3 +23,10 @@ export interface RequestOptions {
   body?: object;
   pathParams?: Record<string, string>;
 }
+
+export interface PaginatedResponse<TItem> {
+  count: number;
+  next: string | null;
+  previous: string | null;
+  results: TItem[];
+}
